requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
opencv-python
numpy
roboflow
